[Puppet] Implemented NDK Crash + GetMinidump directory exposed

+++ build.cake
@@ -284,9 +284,19 @@ Task("BuildAndroidContentProvider").Does(()=>
     BuildAndroidLibrary(appName, libraryName);
     libraryName = "appcenter-push-delegate";
     BuildAndroidLibrary(appName, libraryName);
+
+    // This is a workaround for NDK build making an error where it claims
+    // it can't find the built .so library (which is built successfully).
+    // This error is breaking the CI build on Windows. If you are seeing this, 
+    // most likely we haven't found a fix and this is an NDK bug.
+    if (!IsRunningOnWindows())
+    {
+        appName = "BreakpadSupport";
+        BuildAndroidLibrary(appName, "", false);
+    }
 }).OnError(HandleError);
 
-void BuildAndroidLibrary(string appName, string libraryName) {
+void BuildAndroidLibrary(string appName, string libraryName, bool copyBinary = true) {
     var libraryFolder = System.IO.Path.Combine(appName, libraryName);
     var gradleScript = System.IO.Path.Combine(libraryFolder, "build.gradle");
 
@@ -299,20 +309,22 @@ void BuildAndroidLibrary(string appName, string libraryName) {
     var fullArgs = "-b " + gradleScript + " assembleRelease";
     StartProcess(gradleWrapper, fullArgs);
 
-    // Source and destination of generated aar
-    var aarName = libraryName + "-release.aar";
-    var aarSource = System.IO.Path.Combine(libraryFolder, "build/outputs/aar/" + aarName);
-    var aarDestination = "Assets/AppCenter/Plugins/Android";
+    if (copyBinary) {
+        // Source and destination of generated aar
+        var aarName = libraryName + "-release.aar";
+        var aarSource = System.IO.Path.Combine(libraryFolder, "build/outputs/aar/" + aarName);
+        var aarDestination = "Assets/AppCenter/Plugins/Android";
 
-    // Delete the aar in case it already exists in the Assets folder
-    var existingAar = System.IO.Path.Combine(aarDestination, aarName);
-    if (FileExists(existingAar))
-    {
-        DeleteFile(existingAar);
+        // Delete the aar in case it already exists in the Assets folder
+        var existingAar = System.IO.Path.Combine(aarDestination, aarName);
+        if (FileExists(existingAar))
+        {
+            DeleteFile(existingAar);
+        }
+
+        // Move the .aar to Assets/AppCenter/Plugins/Android
+        MoveFileToDirectory(aarSource, aarDestination);
     }
-
-    // Move the .aar to Assets/AppCenter/Plugins/Android
-    MoveFileToDirectory(aarSource, aarDestination);
 }
 
 // Install Unity Editor for Windows
